frmts/sdts: fix MSVC warnings about conversion size_t->int and forcing int to bool
git-svn-id: https://svn.osgeo.org/gdal/trunk@31596 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/sdts/sdtscatd.cpp
@@ -121,7 +121,7 @@ int SDTS_CATD::Read( const char * pszFilename )
 /*      Strip off the filename, and keep the path prefix.               */
 /* -------------------------------------------------------------------- */
     pszPrefixPath = CPLStrdup( pszFilename );
-    int i = strlen(pszPrefixPath) - 1;
+    int i = static_cast<int>(strlen(pszPrefixPath)) - 1;
     for( ; i > 0; i-- )
     {
         if( pszPrefixPath[i] == '\\' || pszPrefixPath[i] == '/' )
